proper type for blocksize
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@674 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ mformat.c
@@ -916,7 +916,7 @@ void mformat(int argc, char **argv, int dummy UNUSEDP)
 	unsigned int argtracks;
 	uint16_t argheads, argsectors;
 	uint32_t tot_sectors=0;
-	size_t blocksize;
+	uint32_t blocksize;
 
 	char drive, name[EXPAND_BUF];
